requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-jwt-extended>=4.6.0
flask-cors>=4.0.0
flask-restx
redis[hiredis]
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings
gunicorn>=21.0.0
psycopg2-binary>=2.9.0
dependency-injector>=4.41.0
passlib
sqlalchemy

# Dev packages (nếu muốn trong dev container)
ruff
mypy
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-mock>=3.12.0
pre-commit>=3.7.0
bandit[toml]
types-redis
types-flask
factory-boy>=3.3.0
